make Database journalable
