+++ core/server/common/src/main/java/alluxio/master/journal/checkpoint/CheckpointName.java
@@ -19,6 +19,7 @@ public enum CheckpointName {
   ACTIVE_SYNC_MANAGER,
   BLOCK_MASTER,
   CACHING_INODE_STORE,
+  CATALOG_SERVICE_DATABASE,
   CATALOG_SERVICE_MASTER,
   CLUSTER_INFO,
   FILE_SYSTEM_MASTER,
